Add verifiers
